fix some api

+++ api/src/food/food.service.ts
@@ -28,9 +28,21 @@ export class FoodService {
       const food = await this.prisma.food.findUnique({
         where: {
           foodId: foodId
+        },
+        select: {
+          foodId: true,
+          name: true,
+          description: true,
+          categories: {
+            select: {
+              category: true
+            }
+          }
         }
       });
       if (food) {
+        food['category'] = food.categories[0].category;
+        delete food.categories;
         return food;
       } else {
         return { message: 'Food not found' };
@@ -61,15 +73,54 @@ export class FoodService {
 
   async remove(id: number) {
     try {
+      const listOfRecipe = await this.prisma.recipeFoodList.findMany({
+        where: { foodId: id }
+      });
+      if (listOfRecipe.length > 0) {
+        return {
+          statusCode: 400,
+          message: 'Thức ăn đã tồn tại trong công thức, không thể xóa'
+        };
+      }
+      const listOfStorage = await this.prisma.storageFood.findMany({
+        where: { foodId: id }
+      });
+      if (listOfStorage.length > 0) {
+        return {
+          statusCode: 400,
+          message: 'Thức ăn đã tồn tại trong kho lưu trữ, không thể xóa'
+        };
+      }
+      const listOfToBuyLlist = await this.prisma.toBuyListDetail.findMany({
+        where: { foodId: id }
+      });
+      if (listOfToBuyLlist.length > 0) {
+        return {
+          statusCode: 400,
+          message:
+            'Thức ăn đã tồn tại trong danh sách đồ cần mua, không thể xóa'
+        };
+      }
+      await this.prisma.foodCategory.deleteMany({
+        where: { foodId: id }
+      });
       const updateFood = await this.prisma.food.delete({
         where: { foodId: id }
       });
-      if (updateFood) return { message: 'Deleted successfully' };
+      if (updateFood) return { message: 'Xóa thành công' };
       else {
-        throw new Error('Not found');
+        return {
+          statusCode: 400,
+          message: 'Không tìm thấy đồ ăn cần xóa'
+        };
       }
     } catch (err) {
-      throw err;
+      if (err.code == 'P2025') {
+        return {
+          statusCode: 400,
+          message: 'Không tìm thấy đồ ăn cần xóa'
+        };
+      }
     }
   }
 }
